Hide when no articles
My Cloudflare Worker has been acting up lately, so until I find a fix I need a fallback for the site so there isn't just an empty section. My current approach is to hide the section and the relevant nav element, but I hope to find a better workaround in the near future so I can re-implement the staggered bolding of nav elements.