Dynamic settings almost finished
There are still a few bugs to resolve, notably the lack of a scroll bar in the settings window when one is necessary, the tendency of new settings to apply alongside the old settings instead of replacing them, and the boolean checkbox refusing to allow interactivity. But everything else is working swimmingly, we are almost there.