Improved matrix wallpaper
The code could still possibly see performance benefits in the future from transitioning to requestAnimationFrame, but for the time being this is going to have to be good enough. I reduced some of the cycles associated with each canvas iteration as well as vastly improved the resize behavior while keeping the smallest possible canvas to cover the viewport.