Transition to JWT auth
The previous method of authentication involved storing the username/password as a cleartext cookie on the client, whereas now a refresh token is stored that can be revoked from the server-side database at any time. This is far from perfect, but it will be a step in the right direction as we explore alternative options moving forwards.

Related to the now-defunct https://github.com/os-js/osjs-server/pull/68.