New We10X icon theme
Still has a LOT of files that could be pruned, as you can see, but for now it works and it's something. This introduces some issues on the client since it uses a lot of SVGs, which has already been observed in https://github.com/os-js/osjs-client/issues/150. Will be looking for a fix and hopefully making a PR soon.