Improved `script` option assignment
Anders pointed out that the `then` chain is only no longer supported for the Google API, but that it will still work fine for the `script` function itself. So I was able to get rid of the `onload` and `onerror` overloading options to ensure that the promise is not broken.

Ref: https://github.com/os-js/osjs-client/pull/193